Move database filtering out of useDataPickerConfig - Reuse DataTypeInfoItem type - Remove Bucket type - Make getDataTypes conditionally return raw data option
