+++ frontend/src/metabase/containers/DataPicker/types.ts
@@ -5,9 +5,10 @@ import type {
   TableId,
 } from "metabase-types/api";
 
+import type { IconName } from "metabase/core/components/Icon";
 import type Database from "metabase-lib/metadata/Database";
-import type Table from "metabase-lib/metadata/Table";
 import type Schema from "metabase-lib/metadata/Schema";
+import type Table from "metabase-lib/metadata/Table";
 
 export type DataPickerDataType = "models" | "raw-data" | "questions";
 
@@ -38,4 +39,11 @@ export interface DataPickerProps {
 export type DataPickerSelectedItem = {
   type: "database" | "schema" | "collection" | "table";
   id: string | number;
+};
+
+export type DataTypeInfoItem = {
+  id: DataPickerDataType;
+  icon: IconName;
+  name: string;
+  description: string;
 };
